src/requirements.txt
beautifulsoup4==4.9.1
bert-extractive-summarizer==0.5.0
blis==0.4.1
Brotli==1.0.9
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cymem==2.0.3
dash==1.15.0
dash-bootstrap-components==0.10.5
dash-core-components==1.11.0
dash-html-components==1.1.0
dash-renderer==1.7.0
dash-table==4.10.0
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
murmurhash==1.0.2
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
preshed==3.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
retrying==1.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sd-material-ui==3.2.0
sentencepiece==0.1.91
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==2.2.3
tokenizers==0.8.1rc1
torch==1.6.0
tqdm==4.48.2
transformers==3.0.2
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.1.0
